Add floating numbers for combos, change rules for building combos
when the player's combo changes, a floating number particle such as
"x2", "x3", etc will appear to provide feedback to the player. Combos
are now only reset when a move is made that does not directly result in
a match forming. Combos are only increased by matches, they are no
longer affected by destroying junk